requirements.txt
openai
openai-whisper
SpeechRecognition
PyAudio
soundfile
numpy
keyboard
pydub